moved function prototypes to header file
